many fixes to simulation script
Now uses FFTVis for diffuse simulation so we don't need to add a
monopole and worry about the short baselines picking some of it up.

Diffuse matrix should be more-or-less at the right amplitude now, at
least within a factor of 2 or so.

Baseline conjugation convention has been fixed to match what is used
by FFTVis/MatVis.

Updates made to be compliant with new versions of pyuvdata and packages
that depend on pyuvdata.