doc: update tutorial notebook
It should now work for current versions of pyuvdata and packages that
rely on pyuvdata; however, this means it is now broken for environments
that use old versions of pyuvdata (i.e., before analytic beams were
removed from pyuvsim and added to pyuvdata). The text for figures should
be updated, and the discussion about the covariance models should be
updated to better track the discussion in the paper. As a future task,
it would probably be a good idea to make a notebook that just goes over
the diffuse matrix calculation to show that the various approximations
made in the paper hold.